+++ app/models/annotation_category.rb
@@ -2,9 +2,9 @@ class AnnotationCategory < ApplicationRecord
   has_many :annotation_texts, dependent: :destroy
 
   validates_presence_of :annotation_category_name
-  validates_uniqueness_of :annotation_category_name, scope: :assignment_id
+  validates_uniqueness_of :annotation_category_name, scope: :assessment_id
 
-  belongs_to :assignment
+  belongs_to :assignment, foreign_key: :assessment_id
 
   # Takes an array of comma separated values, and tries to assemble an
   # Annotation Category, and associated Annotation Texts