+++ config/initializers/inflections.rb
@@ -18,4 +18,5 @@
 ActiveSupport::Inflector.inflections(:en) do |inflect|
   inflect.irregular 'criterion', 'criteria'
   inflect.irregular 'ta', 'tas'
+  inflect.uncountable 'AssignmentProperties'
 end