+++ db/seeds.rb
@@ -5,6 +5,9 @@ else
   FileUtils.mkdir_p(Rails.configuration.x.repository.storage)
 end
 
+FileUtils.mkdir_p('tmp')
+
+
 # run tasks
 Rake::Task['db:admin'].invoke
 Rake::Task['db:tas'].invoke
@@ -19,5 +22,3 @@ Rake::Task['db:remarks'].invoke
 Rake::Task['db:peer_reviews'].invoke
 Rake::Task['db:scanned_exam'].invoke
 Rake::Task['db:marking_scheme'].invoke
-
-Rake::Task['db:autotest'].invoke unless ENV['SEED_AUTOTEST'] == 'false'