+++ spec/controllers/groups_controller_spec.rb
@@ -152,10 +152,7 @@ describe GroupsController do
         # Setup for Git Repository
         allow(Rails.configuration.x.repository).to receive(:type).and_return('git')
 
-        @assignment = create(:assignment,
-                             allow_web_submits: true,
-                             group_max: 1,
-                             group_min: 1)
+        @assignment = create(:assignment)
 
         # Create students corresponding to the file_good
         @student_user_names = %w(c8shosta c5bennet)
@@ -203,10 +200,7 @@ describe GroupsController do
       end
 
       before :each do
-        @assignment = FactoryBot.create(:assignment,
-                                         allow_web_submits: true,
-                                         group_max: 1,
-                                         group_min: 1)
+        @assignment = FactoryBot.create(:assignment)
 
         @group = FactoryBot.create(:group)
 
@@ -218,13 +212,6 @@ describe GroupsController do
 
         grouping.add_member(@student1, StudentMembership::STATUSES[:inviter])
         grouping.add_member(@student2, StudentMembership::STATUSES[:accepted])
-
-        @ta_name = 'c8shacd'
-        @ta = create(:ta, user_name: @ta_name)
-        # For each grouping for Assignment 1, assign 2 TAs
-        @assignment.groupings.each do |grouping|
-          grouping.add_tas([@ta])
-        end
       end
 
       it 'responds with appropriate status' do
@@ -305,12 +292,14 @@ describe GroupsController do
   describe 'student access' do
     before :each do
       # Authenticate user is not timed out, and has administrator rights.
+      @current_student = create(:student, user_name: 'c9test2')
       allow(controller).to receive(:session_expired?).and_return(false)
       allow(controller).to receive(:logged_in?).and_return(true)
-      allow(controller).to receive(:current_user).and_return(build(:student))
-
+      allow(controller).to receive(:current_user).and_return(@current_student)
       @student = create(:student, user_name: 'c9test1')
-      @assignment = create(:assignment, student_form_groups: true)
+      @assignment = create(:assignment,
+                           due_date: 1.day.from_now,
+                           assignment_properties_attributes: { student_form_groups: true, group_max: 3 })
     end
 
     describe 'POST #create' do
@@ -330,6 +319,19 @@ describe GroupsController do
 
       it 'should respond with success' do
         is_expected.to respond_with(:redirect)
+      end
+    end
+
+    describe 'POST #invite_member' do
+      before :each do
+        create(:grouping_with_inviter, assignment: @assignment, inviter: @current_student)
+      end
+      it 'should send an email to every student invited to a grouping' do
+        expect do
+          post :invite_member,
+               params: { invite_member: @student.user_name,
+                         assignment_id: @assignment.id }
+        end.to change { ActionMailer::Base.deliveries.count }.by(1)
       end
     end
   end