+++ spec/factories/annotation_texts.rb
@@ -3,7 +3,7 @@ FactoryBot.define do
     content { Faker::Lorem.sentence }
     created_at { Time.now }
     updated_at { Time.now }
-    user { FactoryBot.create(:admin) }
+    creator { FactoryBot.create(:admin) }
     association :annotation_category, factory: :annotation_category
   end
 end