+++ spec/models/rubric_criterion_spec.rb
@@ -46,63 +46,6 @@ describe RubricCriterion do
     end
   end
 
-  context 'A rubric criterion assigning a TA' do
-    before(:each) do
-      @criterion = create(:rubric_criterion)
-      @ta = create(:ta)
-    end
-
-    it 'not assign the same TA multiple times' do
-      expect(@criterion.criterion_ta_associations.count).to eq(0), 'Got unexpected TA membership count'
-      @criterion.add_tas(@ta)
-      @ta.reload
-      @criterion.add_tas(@ta)
-      expect(@criterion.criterion_ta_associations.count).to eq(1), 'Got unexpected TA membership count'
-    end
-
-    it 'unassign a TA by id' do
-      expect(@criterion.criterion_ta_associations.count).to eq(0), 'Got unexpected TA membership count'
-      @criterion.add_tas(@ta)
-      expect(@criterion.criterion_ta_associations.count).to eq(1), 'Got unexpected TA membership count'
-      @ta.reload
-      @criterion.remove_tas(@ta)
-      expect(@criterion.criterion_ta_associations.count).to eq(0), 'Got unexpected TA membership count'
-    end
-
-    it 'assign multiple TAs' do
-      ta1 = create(:ta)
-      ta2 = create(:ta)
-      ta3 = create(:ta)
-      expect(@criterion.criterion_ta_associations.count).to eq(0), 'Got unexpected TA membership count'
-      @criterion.add_tas([ta1, ta2, ta3])
-      expect(@criterion.criterion_ta_associations.count).to eq(3), 'Got unexpected TA membership count'
-    end
-
-    it 'remove multiple TAs' do
-      ta1 = create(:ta)
-      ta2 = create(:ta)
-      ta3 = create(:ta)
-      expect(@criterion.criterion_ta_associations.count).to eq(0)
-      @criterion.add_tas([ta1, ta2, ta3])
-      expect(@criterion.criterion_ta_associations.count).to eq(3)
-      ta1.reload
-      ta2.reload
-      ta3.reload
-      @criterion.remove_tas([ta1, ta3])
-      expect(@criterion.criterion_ta_associations.count).to eq(1)
-      @criterion.reload
-      expect(ta2.id).to be == @criterion.tas[0].id
-    end
-
-    it 'get the names of TAs assigned to it' do
-      ta1 = create(:ta)
-      ta2 = create(:ta)
-      @criterion.add_tas(ta1)
-      @criterion.add_tas(ta2)
-      expect(@criterion.get_ta_names).to contain_exactly(ta1.user_name, ta2.user_name)
-    end
-  end
-
   context 'from an assignment without criteria' do
     before(:each) do
       @assignment = create(:assignment)